+++ src/main/java/com/even/zaro/dto/profile/FollowerFollowingListDto.java
@@ -1,4 +1,4 @@
-package com.even.zaro.dto;
+package com.even.zaro.dto.profile;
 
 public class FollowerFollowingListDto {
 }
